requirements.txt
rich
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
patool
